Use uint32_t instead of size_t in IMessageSocket interface
uint32_t should guarantee a better compatiblity between x64, x86 and arduino architectures

+++ src/communication_interface.h
@@ -14,8 +14,8 @@ class IMessageSocket
 public:
   virtual bool connect() = 0;
   virtual void disconnect() = 0;
-  virtual bool send(const void* data, size_t bytes) = 0;
-  virtual size_t receive(void* data) = 0;
+  virtual bool send(const void* data, uint32_t bytes) = 0;
+  virtual uint32_t receive(void* data) = 0;
 };
 
 enum class ConnectionStatus {
